+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
+- Add CocoTB integration (discover and run tests)
 - Adds snippets 
 - Prevent the Diplomat view from being visible if the extension if not activated.
+
 
 # 0.1.0
 Considered as the first working distributed version.